правки work territories и др

+++ app/src/main/java/ru/practicum/android/diploma/favorites/presentation/FavoritesViewModel.kt
@@ -60,6 +60,6 @@ class FavoritesViewModel(private val favoritesInteractor: FavoritesInteractor) :
     }
 
     private fun onClickDebounce(item: VacancyShort) {
-        screenStateLiveData.postValue(ShortVacancyListUiState.AnyItem(item.vacancyId.toInt()))
+        screenStateLiveData.postValue(ShortVacancyListUiState.AnyItem(item.vacancyId))
     }
 }
